feat: move database driver into context (#19)
* refactor: separate the context out

* refactor: move database driver into context
