+++ graphql-gateway/js-graphql/src/resolvers/user/register.ts
@@ -1,16 +1,11 @@
 import { MutationRegisterArgs } from '../../gql.types';
-import neo4j from 'neo4j-driver';
 import * as dotenv from 'dotenv';
+
 dotenv.config();
 
 const bcrypt = require('bcrypt');
 
-const driver = neo4j.driver(
-  process.env.DB_URL ?? '',
-  neo4j.auth.basic(process.env.DB_USER ?? '', process.env.DB_PASSWORD ?? ''),
-);
-
-const userRegisterResolver = async (_p: any, { input }: MutationRegisterArgs) => {
+const userRegisterResolver = async (_p: any, { input }: MutationRegisterArgs, { driver }:any) => {
     const session = driver.session({ database: 'neo4j' });
     try {
         const { username, email, password } = input;
