+++ graphql-gateway/js-graphql/src/resolvers/user/user-impl.ts
@@ -1,14 +1,8 @@
-import neo4j from 'neo4j-driver';
 import * as dotenv from 'dotenv';
 
 dotenv.config();
 
-const driver = neo4j.driver(
-  process.env.DB_URL ?? '',
-  neo4j.auth.basic(process.env.DB_USER ?? '', process.env.DB_PASSWORD ?? ''),
-);
-
-const userResolver = async (_p: any, _a: any, { me }: any) => {
+const userResolver = async (_p: any, _a: any, { me, driver }: any) => {
     if(!me) throw new Error('Please login first');
     console.log(me.id);
     console.log(typeof(me.id));
