+++ graphql-gateway/js-graphql/src/resolvers/user/user-search.ts
@@ -1,15 +1,9 @@
 import { QueryFindUserArgs, User } from '../../gql.types';
-import neo4j from 'neo4j-driver';
 import * as dotenv from 'dotenv';
 
 dotenv.config();
 
-const driver = neo4j.driver(
-  process.env.DB_URL ?? '',
-  neo4j.auth.basic(process.env.DB_USER ?? '', process.env.DB_PASSWORD ?? ''),
-);
-
-const findUser = async (_p: any, { input }: QueryFindUserArgs): Promise<User | null> => {
+const findUser = async (_p: any, { input }: QueryFindUserArgs, { driver }: any): Promise<User | null> => {
   const session = driver.session({ database: 'neo4j' });
   try {
     const { id, email, username } = input;
@@ -17,13 +11,13 @@ const findUser = async (_p: any, { input }: QueryFindUserArgs): Promise<User | n
     let result: any;
     if (id) {
       query = 'MATCH (u:User {id: $id})-[:HAS]->(uInfo) RETURN u, uInfo';
-      result = await session.executeRead((tx) => tx.run(query, { id }));
+      result = await session.run(query, { id });
     } else if (email) {
       query = 'MATCH (u:User {email: $email})-[:HAS]->(uInfo) = $email RETURN u, uInfo';
-      result = await session.executeRead((tx) => tx.run(query, { email }));
+      result = await session.run(query, { email });
     } else if (username) {
       query = 'MATCH (u:User {username: $username})-[:HAS]->(uInfo) RETURN u, uInfo';
-      result = await session.executeRead((tx) => tx.run(query, { username }));
+      result = await session.run(query, { username });
     } else {
       throw new Error('Please input ID/email/username');
     }
